refactoring collections
